+++ src/app/clients/page.tsx
@@ -1,0 +1,5 @@
+import { ClientList } from "./client-list";
+
+export default function ClientsPage() {
+  return <ClientList />;
+}
