+++ src/lib/store.ts
@@ -1,11 +1,16 @@
 import type { Action, ThunkAction } from "@reduxjs/toolkit";
 import { combineSlices, configureStore } from "@reduxjs/toolkit";
+import { clientsApiSlice } from "./features/clients/clients-api-slice";
 import { counterSlice } from "./features/counter/counter-slice";
 import { quotesApiSlice } from "./features/quotes/quotes-api-slice";
 
 // `combineSlices` automatically combines the reducers using
 // their `reducerPath`s, therefore we no longer need to call `combineReducers`.
-const rootReducer = combineSlices(counterSlice, quotesApiSlice);
+const rootReducer = combineSlices(
+  clientsApiSlice,
+  counterSlice,
+  quotesApiSlice
+);
 // Infer the `RootState` type from the root reducer
 export type RootState = ReturnType<typeof rootReducer>;
 
@@ -19,7 +24,9 @@ export const makeStore = () => {
     // Adding the api middleware enables caching, invalidation, polling,
     // and other useful features of `rtk-query`.
     middleware: (getDefaultMiddleware) => {
-      return getDefaultMiddleware().concat(quotesApiSlice.middleware);
+      return getDefaultMiddleware()
+        .concat(clientsApiSlice.middleware)
+        .concat(quotesApiSlice.middleware);
     },
   });
 };
